+++ changelog.md
@@ -1,13 +1,17 @@
 ## 0.4.1
+- `OptionsMarshall` unsafe creation API
+  - `Make{Option|Result}SuccessUnsafe`
+  - `Make{Result|ErrorState}ErrorUnsafe`
 - add `ValuesIntoOrFirstError` (#45)
 - loosen `ValuesInto...` linq extensions from `IList<T>` to `ICollection<T>`
+- add `Match` for `Option<(T1, T2)>` and `Result<(T1, T2){, T}>` with a nicer delegate (for up to 5 elements in the tuple) (#37)
 - updated TUnit to 1.33.0
 
 ## 0.4.0
 - drop .NET 9
 - add `Join` for nullable types
 - add `Join` for 5 elements
-- add `Map` and `Consume` for `Option<(T1, T2)>` and `Result<(T1, T2){, T}>` with a nicer delegate (for up to 5 elements in the tuple)
+- add `Map` and `Consume` for `Option<(T1, T2)>` and `Result<(T1, T2){, T}>` with a nicer delegate (for up to 5 elements in the tuple) (#37)
   - if this concept works more operations will follow in the 0.4 release cycle
 - remove obsolete `OptionsMarshall.TryGetValue` and `TryGetError` (use `Branch` #31)
 - mark tuple operations obsolete (use `Join` #37)